Time: 0 ms (100.00%) | Memory: 42.4 MB (79.67%) - LeetSync

+++ 2619-categorize-box-according-to-criteria/categorize-box-according-to-criteria.java
@@ -1,13 +1,37 @@
+// class Solution {
+//     public String categorizeBox(long length, long width, long height, long mass) {
+//         long volume=(long)length*width*height;
+//         if((length>=10000||width>=10000||height>=10000||volume>=1000000000L) && (mass>=100))
+//    // if(volume>=1000000000&&mass>=100)
+//         {
+//             return "Both";
+//         }
+       
+//         else if(length>=10000||width>=10000||height>=10000||volume>=1000000000L)
+//         {
+//             return "Bulky";
+//         }
+//          else if(mass>=100)
+//         {
+//             return "Heavy";
+//         }
+//         else
+//         {
+//             return "Neither";
+//         }
+//         //return null;
+//     }
+// }
+
 class Solution {
-    public String categorizeBox(long length, long width, long height, long mass) {
+    public String categorizeBox(int length, int width, int height, int mass) {
         long volume=(long)length*width*height;
-        if((length>=10000||width>=10000||height>=10000||volume>=1000000000L) && (mass>=100))
-   // if(volume>=1000000000&&mass>=100)
+        if((length>=10000||width>=10000||height>=10000||volume>=1000000000) && (mass>=100))
         {
             return "Both";
         }
        
-        else if(length>=10000||width>=10000||height>=10000||volume>=1000000000L)
+        else if(length>=10000||width>=10000||height>=10000||volume>=1000000000)
         {
             return "Bulky";
         }
